+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - [+] add `test()` and `replace()` functions
 - [+] add `slice()` and `slice2()` functions
 - [+] add `indexIn()` and `existsIn()` functions
+- [*] `isEmpty()` function can check object for emptiness
 
 
 ## 0.1.0 (10.01.2018)